fix benchmark buildscript
