+++ benchmark/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-  id("flare.base-conventions")
+  id("flare.common-conventions")
   id("me.champeau.jmh") version "0.6.5"
 }
